test: change extra axes to take a percentage as well
And change them to doubles, we need more granularity when the range is only
0-100.

+++ test/litest.h
@@ -195,7 +195,7 @@ struct litest_device {
 
 struct axis_replacement {
 	int32_t evcode;
-	int32_t value;
+	double value;
 };
 
 /* A loop range, resolves to:
